changed drawing order in assignLetter + alphabetView
in order to get rid of the problem when tabbing a letter that has
already been replaced by a photo > didn’t react immediately then. now
I’m drawing the grey grid on top of the letter images. it at the same
time preserves the grey grid even when all letters are replaced by
photos.